feat: cdc checker (#460)
* refactor: split checker and sinker

* feat: wire cdc check runner

* refactor: remove checker arc

* feat: add s3 check logs

* feat: persist checker state

* test: refresh cdc summaries

* refactor(check): simplify checker recovery state

* fix cdc checker flow and simplify dt-tests

* trim checker hot-path clones

* refactor: stage checker simplification baseline

* refactor: simplify checker state scoping

* refactor: split task kind and check mode

* fix: wire standalone struct checker

* refactor: share cdc+check checkpoint state

* test(task-id): lock check_log default task id

Add a canary test that freezes the default task_id for check_log configs so internal enum churn does not silently change external prefixes and orchestration contracts.

* refactor(parallelizer): inline merge sink dispatch

Inline the merge sink dispatch into MergeParallelizer::sink_dml and drop the extra helper layer to keep the hot path simpler.

* fix(checker): restore critical follow-up fixes

* cleanup some tests

* check cdc resume

* feat: add refresh_meta to update metadata in checkers

* simplify check configurations

* enhance test configurations and cleanup

* fix docs and add more tests

* refactor: update checker methods to public and simplify test configurations

* enhance checker configuration and support for CDC checks & cleanup some deprecated codes

* simplify StructCheckerHandle

* simplify

* fix lint and simplify

* simplify and limit sql.log

* fix docs, config, tests

* no fast-fail async check sidecar

* fix task configuration and error handling & enhance mongo extractor

* chore

* cdc check manage persisted ids

* cleanup

* cleanup

* remove some deep copy & add new test cases for cdc resume

* enforce parallelizer for CDC tasks & cleanup

* chore

* chore: cdc check no throw up error & eviction check queue & only persistent pk for unconsistent_rows

* use rdb_merge instead of rdb_check

* fix import order

* feat: use tagged_col_value_map manage checker state

* rename to checkable_sinker

* enhance sql log and check state store

* refactor: improve inline CDC check logging and add local benchmark reference

* cleanup

* docs

* update docs

* revert

* cleanup

* chore: remove unnecessary clone

---------

+++ dt-tests/tests/test_runner/redis_test_runner.rs
@@ -1,6 +1,3 @@
-use crate::test_runner::redis_test_util::RedisTestUtil;
-
-use super::{base_test_runner::BaseTestRunner, redis_cluster_connection::RedisClusterConnection};
 use anyhow::bail;
 use dt_common::{
     config::{
@@ -16,6 +13,9 @@ use dt_common::{
 };
 
 use redis::{Connection, Value};
+
+use super::{base_test_runner::BaseTestRunner, redis_cluster_connection::RedisClusterConnection};
+use crate::test_runner::redis_test_util::RedisTestUtil;
 
 pub struct RedisTestRunner {
     pub base: BaseTestRunner,
@@ -421,13 +421,13 @@ impl RedisTestRunner {
         let (src_to_compare, dst_to_compare) = if let Some(idx) = result_index {
             match (&src_result, &dst_result) {
                 (Value::Array(src_arr), Value::Array(dst_arr)) => {
-                    let src_elem = src_arr.get(idx as usize).unwrap_or(&Value::Nil);
-                    let dst_elem = dst_arr.get(idx as usize).unwrap_or(&Value::Nil);
+                    let src_elem = src_arr.get(idx).unwrap_or(&Value::Nil);
+                    let dst_elem = dst_arr.get(idx).unwrap_or(&Value::Nil);
                     (src_elem, dst_elem)
                 }
                 (Value::Set(src_set), Value::Set(dst_set)) => {
-                    let src_elem = src_set.get(idx as usize).unwrap_or(&Value::Nil);
-                    let dst_elem = dst_set.get(idx as usize).unwrap_or(&Value::Nil);
+                    let src_elem = src_set.get(idx).unwrap_or(&Value::Nil);
+                    let dst_elem = dst_set.get(idx).unwrap_or(&Value::Nil);
                     (src_elem, dst_elem)
                 }
                 _ => (&src_result, &dst_result),
